requirements.txt
rich>=13.0
Pillow>=10.0
